+++ src/vmware/azext_vmware/custom.py
@@ -21,6 +21,22 @@ VMWARE DATA PROCESSING AGREEMENT. Once Professional Services Data is transferred
 ACCEPTANCE OF LEGAL TERMS. By continuing, you agree to the above additional Legal Terms for AVS. If you are an individual accepting these terms on behalf of an entity, you also represent that you have the legal authority to enter into these additional terms on that entity's behalf.
 '''
 
+ROTATE_VCENTER_PASSWORD_TERMS = '''
+Any services connected using these credentials will stop working and may cause you to be locked out of your account.
+
+Check if you're using your cloudadmin credentials for any connected services like backup and disaster recovery appliances, VMware HCX, or any vRealize suite products. Verify you're not using cloudadmin credentials for connected services before generating a new password.
+
+If you are using cloudadmin for connected services, learn how you can setup a connection to an external identity source to create and manage new credentials for your connected services: https://docs.microsoft.com/en-us/azure/azure-vmware/configure-identity-source-vcenter
+
+Press Y to confirm no services are using my cloudadmin credentials to connect to vCenter
+'''
+
+ROTATE_NSXT_PASSWORD_TERMS = '''
+Currently, rotating your NSX-T managed admin credentials isn’t supported.  If you need to rotate your NSX-T manager admin credentials, please submit a support request in the Azure Portal: https://portal.azure.com/#create/Microsoft.Support
+
+Press any key to continue
+'''
+
 
 def privatecloud_list(client: AVSClient, resource_group_name=None):
     if resource_group_name is None:
@@ -32,12 +48,12 @@ def privatecloud_show(client: AVSClient, resource_group_name, name):
     return client.private_clouds.get(resource_group_name, name)
 
 
-def privatecloud_create(client: AVSClient, resource_group_name, name, sku, cluster_size, network_block, location=None, internet=None, vcenter_password=None, nsxt_password=None, tags=None, accept_eula=False, mi_system_assigned=False):
+def privatecloud_create(client: AVSClient, resource_group_name, name, sku, cluster_size, network_block, location=None, internet=None, vcenter_password=None, nsxt_password=None, tags=None, accept_eula=False, mi_system_assigned=False, yes=False):
     from knack.prompting import prompt_y_n
     if not accept_eula:
         print(LEGAL_TERMS)
         msg = 'Do you agree to the above additional terms for AVS?'
-        if not prompt_y_n(msg, default="n"):
+        if not yes and not prompt_y_n(msg, default="n"):
             return None
 
     from azext_vmware.vendored_sdks.avs_client.models import PrivateCloud, Circuit, ManagementCluster, Sku, PrivateCloudIdentity
@@ -91,7 +107,11 @@ def privatecloud_addidentitysource(client: AVSClient, resource_group_name, name,
     return client.private_clouds.begin_create_or_update(resource_group_name=resource_group_name, private_cloud_name=private_cloud, private_cloud=pc)
 
 
-def privatecloud_deleteidentitysource(client: AVSClient, resource_group_name, name, private_cloud, alias, domain):
+def privatecloud_deleteidentitysource(client: AVSClient, resource_group_name, name, private_cloud, alias, domain, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the identity source. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     pc = client.private_clouds.get(resource_group_name, private_cloud)
     found = next((ids for ids in pc.identity_sources
                  if ids.name == name and ids.alias == alias and ids.domain == domain), None)
@@ -108,7 +128,11 @@ def privatecloud_addavailabilityzone(client: AVSClient, resource_group_name, pri
     return client.private_clouds.begin_update(resource_group_name=resource_group_name, private_cloud_name=private_cloud, private_cloud_update=pc)
 
 
-def privatecloud_deleteavailabilityzone(client: AVSClient, resource_group_name, private_cloud):
+def privatecloud_deleteavailabilityzone(client: AVSClient, resource_group_name, private_cloud, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the availability zone. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     from azext_vmware.vendored_sdks.avs_client.models import PrivateCloudUpdate
     pc = PrivateCloudUpdate()
     pc.availability = None
@@ -122,7 +146,11 @@ def privatecloud_addcmkencryption(client: AVSClient, resource_group_name, privat
     return client.private_clouds.begin_update(resource_group_name=resource_group_name, private_cloud_name=private_cloud, private_cloud_update=pc)
 
 
-def privatecloud_deletecmkenryption(client: AVSClient, resource_group_name, private_cloud):
+def privatecloud_deletecmkenryption(client: AVSClient, resource_group_name, private_cloud, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the managed keys encryption. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     from azext_vmware.vendored_sdks.avs_client.models import PrivateCloudUpdate
     pc = PrivateCloudUpdate()
     pc.encryption = None
@@ -140,7 +168,7 @@ def privatecloud_identity_assign(client: AVSClient, resource_group_name, private
 def privatecloud_identity_remove(client: AVSClient, resource_group_name, private_cloud):
     from azext_vmware.vendored_sdks.avs_client.models import PrivateCloudIdentity, PrivateCloudUpdate
     pc = PrivateCloudUpdate()
-    pc.identity = PrivateCloudIdentity(type=None)
+    pc.identity = PrivateCloudIdentity(type="None")
     return client.private_clouds.begin_update(resource_group_name=resource_group_name, private_cloud_name=private_cloud, private_cloud_update=pc)
 
 
@@ -148,12 +176,19 @@ def privatecloud_identity_get(client: AVSClient, resource_group_name, private_cl
     return client.private_clouds.get(resource_group_name, private_cloud).identity
 
 
-def privatecloud_rotate_vcenter_password(client: AVSClient, resource_group_name, private_cloud):
+def privatecloud_rotate_vcenter_password(client: AVSClient, resource_group_name, private_cloud, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = ROTATE_VCENTER_PASSWORD_TERMS
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.private_clouds.begin_rotate_vcenter_password(resource_group_name=resource_group_name, private_cloud_name=private_cloud)
 
 
-def privatecloud_rotate_nsxt_password(client: AVSClient, resource_group_name, private_cloud):
-    return client.private_clouds.begin_rotate_nsxt_password(resource_group_name=resource_group_name, private_cloud_name=private_cloud)
+def privatecloud_rotate_nsxt_password():
+    from knack.prompting import prompt
+    msg = ROTATE_NSXT_PASSWORD_TERMS
+    prompt(msg)
+    # return client.private_clouds.begin_rotate_nsxt_password(resource_group_name=resource_group_name, private_cloud_name=private_cloud)
 
 
 def cluster_create(client: AVSClient, resource_group_name, name, sku, private_cloud, size, hosts):
@@ -173,7 +208,11 @@ def cluster_show(client: AVSClient, resource_group_name, private_cloud, name):
     return client.clusters.get(resource_group_name=resource_group_name, private_cloud_name=private_cloud, cluster_name=name)
 
 
-def cluster_delete(client: AVSClient, resource_group_name, private_cloud, name):
+def cluster_delete(client: AVSClient, resource_group_name, private_cloud, name, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the cluster. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.clusters.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, cluster_name=name)
 
 
@@ -197,7 +236,11 @@ def authorization_show(client: AVSClient, resource_group_name, private_cloud, na
     return client.authorizations.get(resource_group_name=resource_group_name, private_cloud_name=private_cloud, authorization_name=name)
 
 
-def authorization_delete(client: AVSClient, resource_group_name, private_cloud, name):
+def authorization_delete(client: AVSClient, resource_group_name, private_cloud, name, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the authorization. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.authorizations.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, authorization_name=name)
 
 
@@ -213,7 +256,11 @@ def hcxenterprisesite_show(client: AVSClient, resource_group_name, private_cloud
     return client.hcx_enterprise_sites.get(resource_group_name=resource_group_name, private_cloud_name=private_cloud, hcx_enterprise_site_name=name)
 
 
-def hcxenterprisesite_delete(client: AVSClient, resource_group_name, private_cloud, name):
+def hcxenterprisesite_delete(client: AVSClient, resource_group_name, private_cloud, name, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the HCX enterprise site. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.hcx_enterprise_sites.delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, hcx_enterprise_site_name=name)
 
 
@@ -241,7 +288,11 @@ def datastore_show(client: AVSClient, resource_group_name, private_cloud, cluste
     return client.datastores.get(resource_group_name=resource_group_name, private_cloud_name=private_cloud, cluster_name=cluster, datastore_name=name)
 
 
-def datastore_delete(client: AVSClient, resource_group_name, private_cloud, cluster, name):
+def datastore_delete(client: AVSClient, resource_group_name, private_cloud, cluster, name, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the datastore. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.datastores.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, cluster_name=cluster, datastore_name=name)
 
 
@@ -297,15 +348,27 @@ def addon_srm_update(client: AVSClient, resource_group_name, private_cloud, lice
     return client.addons.begin_create_or_update(resource_group_name=resource_group_name, private_cloud_name=private_cloud, addon_name="srm", properties=properties)
 
 
-def addon_vr_delete(client: AVSClient, resource_group_name, private_cloud):
+def addon_vr_delete(client: AVSClient, resource_group_name, private_cloud, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the VR addon. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.addons.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, addon_name="vr")
 
 
-def addon_hcx_delete(client: AVSClient, resource_group_name, private_cloud):
+def addon_hcx_delete(client: AVSClient, resource_group_name, private_cloud, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the HCX addon. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.addons.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, addon_name="hcx")
 
 
-def addon_srm_delete(client: AVSClient, resource_group_name, private_cloud):
+def addon_srm_delete(client: AVSClient, resource_group_name, private_cloud, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the SRM addon. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.addons.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, addon_name="srm")
 
 
@@ -322,7 +385,11 @@ def globalreachconnection_show(client: AVSClient, resource_group_name, private_c
     return client.global_reach_connections.get(resource_group_name=resource_group_name, private_cloud_name=private_cloud, global_reach_connection_name=name)
 
 
-def globalreachconnection_delete(client: AVSClient, resource_group_name, private_cloud, name):
+def globalreachconnection_delete(client: AVSClient, resource_group_name, private_cloud, name, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the global reach connection. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.global_reach_connections.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, global_reach_connection_name=name)
 
 
@@ -338,7 +405,11 @@ def cloud_link_show(client: AVSClient, resource_group_name, private_cloud, name)
     return client.cloud_links.get(resource_group_name=resource_group_name, private_cloud_name=private_cloud, cloud_link_name=name)
 
 
-def cloud_link_delete(client: AVSClient, resource_group_name, private_cloud, name):
+def cloud_link_delete(client: AVSClient, resource_group_name, private_cloud, name, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the cloud link. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.cloud_links.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, cloud_link_name=name)
 
 
@@ -374,7 +445,11 @@ def script_execution_show(client: AVSClient, resource_group_name, private_cloud,
     return client.script_executions.get(resource_group_name=resource_group_name, private_cloud_name=private_cloud, script_execution_name=name)
 
 
-def script_execution_delete(client: AVSClient, resource_group_name, private_cloud, name):
+def script_execution_delete(client: AVSClient, resource_group_name, private_cloud, name, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the script execution. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.script_executions.begin_delete(resource_group_name=resource_group_name, private_cloud_name=private_cloud, script_execution_name=name)
 
 
@@ -406,7 +481,11 @@ def workload_network_dhcp_relay_update(client: AVSClient, resource_group_name, p
     return client.workload_networks.begin_update_dhcp(resource_group_name=resource_group_name, private_cloud_name=private_cloud, dhcp_id=dhcp, properties=properties)
 
 
-def workload_network_dhcp_delete(client: AVSClient, resource_group_name, private_cloud, dhcp: str):
+def workload_network_dhcp_delete(client: AVSClient, resource_group_name, private_cloud, dhcp: str, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the workload network DHCP. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.workload_networks.begin_delete_dhcp(resource_group_name=resource_group_name, private_cloud_name=private_cloud, dhcp_id=dhcp)
 
 
@@ -438,7 +517,11 @@ def workload_network_dns_services_update(client: AVSClient, resource_group_name,
     return client.workload_networks.begin_update_dns_service(resource_group_name=resource_group_name, private_cloud_name=private_cloud, dns_service_id=dns_service, workload_network_dns_service=prop)
 
 
-def workload_network_dns_services_delete(client: AVSClient, resource_group_name, private_cloud, dns_service):
+def workload_network_dns_services_delete(client: AVSClient, resource_group_name, private_cloud, dns_service, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the workload network DNS services. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.workload_networks.begin_delete_dns_service(resource_group_name=resource_group_name, private_cloud_name=private_cloud, dns_service_id=dns_service)
 
 
@@ -462,7 +545,11 @@ def workload_network_dns_zone_update(client: AVSClient, resource_group_name, pri
     return client.workload_networks.begin_update_dns_zone(resource_group_name=resource_group_name, private_cloud_name=private_cloud, dns_zone_id=dns_zone, workload_network_dns_zone=prop)
 
 
-def workload_network_dns_zone_delete(client: AVSClient, resource_group_name, private_cloud, dns_zone):
+def workload_network_dns_zone_delete(client: AVSClient, resource_group_name, private_cloud, dns_zone, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the workload network DNS zone. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.workload_networks.begin_delete_dns_zone(resource_group_name=resource_group_name, private_cloud_name=private_cloud, dns_zone_id=dns_zone)
 
 
@@ -486,7 +573,11 @@ def workload_network_port_mirroring_update(client: AVSClient, resource_group_nam
     return client.workload_networks.begin_update_port_mirroring(resource_group_name=resource_group_name, private_cloud_name=private_cloud, port_mirroring_id=port_mirroring, workload_network_port_mirroring=prop)
 
 
-def workload_network_port_mirroring_delete(client: AVSClient, resource_group_name, private_cloud, port_mirroring):
+def workload_network_port_mirroring_delete(client: AVSClient, resource_group_name, private_cloud, port_mirroring, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the workload network port mirroring. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.workload_networks.begin_delete_port_mirroring(resource_group_name=resource_group_name, private_cloud_name=private_cloud, port_mirroring_id=port_mirroring)
 
 
@@ -518,7 +609,11 @@ def workload_network_segment_update(client: AVSClient, resource_group_name, priv
     return client.workload_networks.begin_update_segments(resource_group_name=resource_group_name, private_cloud_name=private_cloud, segment_id=segment, workload_network_segment=segmentObj)
 
 
-def workload_network_segment_delete(client: AVSClient, resource_group_name, private_cloud, segment):
+def workload_network_segment_delete(client: AVSClient, resource_group_name, private_cloud, segment, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the workload network segment. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.workload_networks.begin_delete_segment(resource_group_name=resource_group_name, private_cloud_name=private_cloud, segment_id=segment)
 
 
@@ -534,7 +629,11 @@ def workload_network_public_ip_create(client: AVSClient, resource_group_name, pr
     return client.workload_networks.begin_create_public_ip(resource_group_name=resource_group_name, private_cloud_name=private_cloud, public_ip_id=public_ip, display_name=display_name, number_of_public_i_ps=number_of_public_ips)
 
 
-def workload_network_public_ip_delete(client: AVSClient, resource_group_name, private_cloud, public_ip):
+def workload_network_public_ip_delete(client: AVSClient, resource_group_name, private_cloud, public_ip, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the workload network public IP. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.workload_networks.begin_delete_public_ip(resource_group_name=resource_group_name, private_cloud_name=private_cloud, public_ip_id=public_ip)
 
 
@@ -558,7 +657,11 @@ def workload_network_vm_group_update(client: AVSClient, resource_group_name, pri
     return client.workload_networks.begin_update_vm_group(resource_group_name=resource_group_name, private_cloud_name=private_cloud, vm_group_id=vm_group, workload_network_vm_group=vmGroup)
 
 
-def workload_network_vm_group_delete(client: AVSClient, resource_group_name, private_cloud, vm_group):
+def workload_network_vm_group_delete(client: AVSClient, resource_group_name, private_cloud, vm_group, yes=False):
+    from knack.prompting import prompt_y_n
+    msg = 'This will delete the workload network VM group. Are you sure?'
+    if not yes and not prompt_y_n(msg, default="n"):
+        return None
     return client.workload_networks.begin_delete_vm_group(resource_group_name=resource_group_name, private_cloud_name=private_cloud, vm_group_id=vm_group)